adds folder for tests
